unfollow user

+++ ressources/js/view/user.js
@@ -48,7 +48,6 @@ define([
                             table.push(userToPrint.toJSON());
                             table.push(getCurrentUser());
                             that.$el.html(that.template({results: table}));
-                            console.log(table);
                         }
                     });
 
@@ -58,35 +57,46 @@ define([
 
         followUser: function(){
             var that = this;
-            var idData = JSON.stringify({id: $(event.target).data('id')});
-            $.ajax({
-                url: URL + '/follow',
-                type: 'POST',
-                data: idData,
-                dataType: "json",
-                contentType: 'application/json'
-            })
-                .done(function(){
-                    $('#followSuccess').fadeIn().delay(5000).fadeOut();
-                    //that.render();
+            var id = $(event.target).data('id');
+            if(id != -1) {
+                var idData = JSON.stringify({id: id});
+                $.ajax({
+                    url: URL + '/follow',
+                    type: 'POST',
+                    data: idData,
+                    dataType: "json",
+                    contentType: 'application/json'
                 })
-                .fail(function(){
-                    $('#followSuccess').fadeIn().delay(5000).fadeOut();
-                })
+                    .done(function () {
+                        $('#followSuccess').fadeIn().delay(5000).fadeOut();
+                        //that.render();
+                    })
+                    .fail(function () {
+                        $('#followError').fadeIn().delay(5000).fadeOut();
+                    })
+            }
+            else{
+                $('#followError').fadeIn().delay(5000).fadeOut();
+            }
         },
 
         unfollowUser: function(){
             var id = $(event.target).data('id');
-            $.ajax({
-                url: URL + '/follow/' + id,
-                type: 'DELETE'
-            })
-                .done(function(){
-                    $('#unfollowSuccess').fadeIn().delay(5000).fadeOut();
+            if (id != -1) {
+                $.ajax({
+                    url: URL + '/follow/' + id,
+                    type: 'DELETE'
                 })
-                .fail(function(){
-                    $('#unfollowError').fadeIn().delay(5000).fadeOut();
-                })
+                    .done(function () {
+                        $('#unfollowSuccess').fadeIn().delay(5000).fadeOut();
+                    })
+                    .fail(function () {
+                        $('#unfollowError').fadeIn().delay(5000).fadeOut();
+                    })
+            }
+            else{
+                $('#unfollowError').fadeIn().delay(5000).fadeOut();
+            }
         }
 
     });
